+++ package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "react": "^16.8.3",
     "react-dom": "^16.8.3",
+    "react-router-dom": "^5.0.0",
     "react-scripts": "2.1.5",
     "react-timeout": "^1.2.0"
   },
